Memory feed is updated.
